// src/features/multistep-forms/forms/claims/burglary/details-of-loss/types/schema.ts
import { z } from "zod";

export const schema = z.object({
  premisesAddress: z.string().min(1, "Required"),
  premisesTelephone: z.string().min(1, "Required"),
  theftDate: z.date(),
  theftTime: z.string().min(1, "Required"),
  entryDetails: z.string().min(1, "Required"),
  roomsEntered: z.string().min(1, "Required"),
  premisesOccupied: z.string(),
  lastOccupiedDateTime: z.string().optional(),
  hasSuspects: z.string(),
  suspectName: z.string().optional(),
  hasInformedPolice: z.string(),
  policeNotificationDate: z.date().optional(),
  policeStationName: z.string().optional(),
  policeStationAddress: z.string().optional(),
  isSoleOwner: z.string(),
  otherOwnersName: z.string().optional(),
  otherOwnersAddress: z.string().optional(),
  hasOtherInsurance: z.string(),
  otherInsurersName: z.string().optional(),
  otherInsurersAddress: z.string().optional(),
  totalContentsValue: z.string().min(1, "Required"),
  firePolicySum: z.string().min(1, "Required"),
  fireInsurersName: z.string().min(1, "Required"),
  fireInsurersAddress: z.string().min(1, "Required"),
  hasPreviousTheft: z.string(),
  previousTheftDetails: z.string().optional(),
  propertyItems: z.array(
    z.object({
      description: z.string().min(1, "Required"),
      purchaseDate: z.date(),
      costPrice: z.string().min(1, "Required"),
      estimatedValue: z.string().min(1, "Required"),
      netAmountClaimed: z.string().min(1, "Required"),
    })
  ),
});

export type Schema = z.infer<typeof schema>;

export const defaultValues: Schema = {
  premisesAddress: "",
  premisesTelephone: "",
  theftDate: new Date(),
  theftTime: "",
  entryDetails: "",
  roomsEntered: "",
  premisesOccupied: "",
  lastOccupiedDateTime: "",
  hasSuspects: "",
  suspectName: "",
  hasInformedPolice: "",
  policeNotificationDate: new Date(),
  policeStationName: "",
  policeStationAddress: "",
  isSoleOwner: "",
  otherOwnersName: "",
  otherOwnersAddress: "",
  hasOtherInsurance: "",
  otherInsurersName: "",
  otherInsurersAddress: "",
  totalContentsValue: "",
  firePolicySum: "",
  fireInsurersName: "",
  fireInsurersAddress: "",
  hasPreviousTheft: "",
  previousTheftDetails: "",
  propertyItems: [],
}; 